+++ dlrover/python/master/diagnosis/diagnosis.py
@@ -13,12 +13,11 @@
 
 import threading
 import time
+from datetime import datetime, timedelta
+from typing import List, Dict
 
+from common.diagnosis import DiagnosisData
 from dlrover.python.common.log import default_logger as logger
-from dlrover.python.master.diagnosis.diagnosis_data import (
-    DataManager,
-    DiagnosisData,
-)
 from dlrover.python.master.diagnosis.diagnostician import Diagnostician
 from dlrover.python.master.diagnosis.inferencechain.common import (
     Inference,
@@ -28,9 +27,15 @@ from dlrover.python.master.diagnosis.inferencechain.common import (
 )
 
 
+def has_expired(timestamp: float, time_period: int) -> bool:
+    dt = datetime.fromtimestamp(timestamp)
+    expired_dt = dt + timedelta(seconds=time_period)
+    return expired_dt < datetime.now()
+
+
 class DiagnosisManager:
     def __init__(self):
-        self.data_manager: DataManager = DataManager(600)
+        self.data_manager: DiagnosisDataManager = DiagnosisDataManager(600)
         self.diagnostician: Diagnostician = Diagnostician(self.data_manager)
 
     def collect_diagnosis_data(self, data_type: str, data: DiagnosisData):
@@ -38,7 +43,7 @@ class DiagnosisManager:
 
     def start(self):
         logger.info("Start Diagnosis Manager ...")
-        problems: list[Inference] = [
+        problems: List[Inference] = [
             Inference(
                 InferenceName.TRAINING,
                 InferenceAttribute.ISORNOT,
@@ -74,3 +79,35 @@ class DiagnosisManager:
                 for root_cause in root_causes:
                     logger.info(f"identify root cause: {root_cause}")
             time.sleep(180)
+
+
+class DiagnosisDataManager:
+    def __init__(self, expire_time_period):
+        self.diagnosis_data: Dict[str, List[DiagnosisData]] = {}
+        self.expire_time_period = expire_time_period
+
+    def store_data(self, data_type: str, data: DiagnosisData):
+        if data_type not in self.diagnosis_data:
+            logger.warning(f"{data_type} is not found in the store")
+            self.diagnosis_data[data_type] = []
+        self.diagnosis_data[data_type].append(data)
+        self._clean_diagnosis_data(data_type)
+
+    def get_data(self, data_type: str) -> List[DiagnosisData]:
+        if data_type not in self.diagnosis_data:
+            return []
+        return self.diagnosis_data[data_type]
+
+    def _clean_diagnosis_data(self, data_type: str):
+        if data_type not in self.diagnosis_data:
+            return
+
+        data = self.diagnosis_data[data_type]
+        n = 0
+        for d in data:
+            if has_expired(d.get_timestamp(), self.expire_time_period):
+                n = n + 1
+            else:
+                break
+
+        self.diagnosis_data[data_type] = data[n:]
