+++ dlrover/python/master/diagnosis/inferencechain/common.py
@@ -15,6 +15,8 @@ from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
 from typing import List
 
+from master.diagnosis.diagnosis import DiagnosisDataManager
+
 
 class InferenceName:
     END = "end"
@@ -45,8 +47,8 @@ class InferenceOperator(metaclass=ABCMeta):
     InferenceOperator is used to infer the root cause of problems.
     """
 
-    def __init__(self):
-        pass
+    def __init__(self, data_manager: DiagnosisDataManager):
+        self._data_manager = data_manager
 
     @abstractmethod
     def infer(self, inferences: List[Inference]) -> List[Inference]:
